Created the second kafka consumer/producer responsible for updating the average

+++ backend-spring-boot/demo/src/main/java/com/example/demo/Model/Student.java
@@ -20,5 +20,5 @@ public class Student {
     private String email;
     private String className;
     private String field;
-    private float average;
+    private double average;
 }
